refactor(collector): extract CheckLoader from CheckScheduler
Move check-loading fields and logic (loaders, senderManager, addLoader,
getChecks) out of CheckScheduler into a new CheckLoader type in
pkg/collector/loader.go.

CheckScheduler now holds a *CheckLoader and delegates instantiation to
CheckLoader.LoadChecks(config, sm). The explicit sm parameter replaces
the implicit field access, enabling callers to route samples to an
alternate SenderManager (e.g. a shadow pipeline) without loading a
separate config.

InitCheckScheduler signature is unchanged — all existing callers
continue to compile without modification.

This is a pure structural refactor with no behaviour change. It lays
the groundwork for the shadow check pipeline (PR #51938), where
collectorimpl will receive *CheckLoader as an optional fx dependency
and call ConfigureShadow on it, breaking the would-be circular
fx dependency CheckScheduler → collector.Component → collectorimpl → CheckScheduler.
